requirements.txt
requests>=2.31.0
sqlalchemy>=2.0.0
pandas>=2.1.0
numpy>=1.25.0
scikit-learn>=1.3.0
lightgbm>=4.1.0
xgboost>=2.0.0
streamlit>=1.29.0
plotly>=5.18.0
pyyaml>=6.0.1
schedule>=1.2.0
tqdm>=4.66.0
